requirements.txt
aniso8601==9.0.1
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
comtypes==1.2.0
Flask==2.3.2
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gTTS==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
pypiwin32==223
python-engineio==4.6.1
python-socketio==5.8.0
pyttsx3==2.90
pytz==2023.3
pywin32==306
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
